convert theming to TS
