+++ package.json
@@ -236,7 +236,7 @@
     "timemachine": "0.3.2",
     "ts-jest": "26.5.6",
     "ts-loader": "9.3.1",
-    "typescript": "4.7.4",
+    "typescript": "4.6.4",
     "wait-on": "6.0.1",
     "webpack": "5.72.1",
     "webpack-bundle-analyzer": "4.5.0",
