--- src/app/theming/index.js
@@ -1,2 +1,0 @@
-export { default as mundo } from './services/mundo';
-export { default as pidgin } from './services/pidgin';
