+++ src/app/theming/index.ts
@@ -1,0 +1,2 @@
+export { default as mundo } from './services/mundo';
+export { default as pidgin } from './services/pidgin';
