--- src/app/theming/services/index.js
@@ -1,2 +1,0 @@
-export { default as mundo } from './mundo';
-export { default as pidgin } from './pidgin';
