+++ src/app/theming/services/index.ts
@@ -1,0 +1,2 @@
+export { default as mundo } from './mundo';
+export { default as pidgin } from './pidgin';
