+++ src/app/theming/services/mundo.ts
@@ -1,0 +1,33 @@
+import { GHOST, WHITE, POSTBOX, POSTBOX_30 } from '../colours';
+import latin from '../typography/scripts/latin';
+import {
+  REITH_SANS_BOLD,
+  REITH_SANS_REGULAR,
+  REITH_SERIF_MEDIUM,
+  REITH_SERIF_LIGHT,
+} from '../typography/fontFaces';
+
+const mundo = {
+  colours: {
+    BRAND_BACKGROUND: POSTBOX,
+    BRAND_LOGO: WHITE,
+    BRAND_FOREGROUND: GHOST,
+    BRAND_HIGHLIGHT: WHITE,
+    BRAND_BORDER: POSTBOX_30,
+  },
+  typography: {
+    script: latin,
+    fontFamilyVariants: {
+      primary: 'ReithSans',
+      secondary: 'ReithSerif',
+    },
+    fontFaces: [
+      REITH_SANS_BOLD,
+      REITH_SANS_REGULAR,
+      REITH_SERIF_MEDIUM,
+      REITH_SERIF_LIGHT,
+    ],
+  },
+};
+
+export default mundo;
