--- src/app/theming/spacings/index.js
@@ -1,6 +1,0 @@
-import pixelsToRem from '../../utilities/pixelsToRem';
-
-export const FULL = pixelsToRem(16);
-export const DOUBLE = FULL * 2;
-export const TRIPLE = FULL * 3;
-export const QUADRIPLE = FULL * 4;
