+++ src/app/theming/spacings/index.ts
@@ -1,0 +1,6 @@
+import pixelsToRem from '../../utilities/pixelsToRem';
+
+export const FULL = pixelsToRem(16);
+export const DOUBLE = FULL * 2;
+export const TRIPLE = FULL * 3;
+export const QUADRIPLE = FULL * 4;
