--- src/app/theming/typography/scripts/cyrillic.js
@@ -1,8 +1,0 @@
-import latin from './latin';
-
-/*
- * The typography values for Latin and Cyrillic characters are identical
- */
-const cyrillic = latin;
-
-export default cyrillic;
