+++ src/app/theming/typography/scripts/cyrillic.ts
@@ -1,0 +1,8 @@
+import latin from './latin';
+
+/*
+ * The typography values for Latin and Cyrillic characters are identical
+ */
+const cyrillic = latin;
+
+export default cyrillic;
