+++ src/app/theming/typography/scripts/latin.ts
@@ -1,0 +1,216 @@
+import pixelsToRem from '../../../utilities/pixelsToRem';
+
+const latin = {
+  atlas: {
+    groupA: {
+      fontSize: pixelsToRem(78),
+      lineHeight: pixelsToRem(84),
+    },
+    groupB: {
+      fontSize: pixelsToRem(96),
+      lineHeight: pixelsToRem(104),
+    },
+    groupD: {
+      fontSize: pixelsToRem(140),
+      lineHeight: pixelsToRem(148),
+    },
+  },
+  elephant: {
+    groupA: {
+      fontSize: pixelsToRem(60),
+      lineHeight: pixelsToRem(64),
+    },
+    groupB: {
+      fontSize: pixelsToRem(78),
+      lineHeight: pixelsToRem(84),
+    },
+    groupD: {
+      fontSize: pixelsToRem(116),
+      lineHeight: pixelsToRem(124),
+    },
+  },
+  imperial: {
+    groupA: {
+      fontSize: pixelsToRem(50),
+      lineHeight: pixelsToRem(54),
+    },
+    groupB: {
+      fontSize: pixelsToRem(64),
+      lineHeight: pixelsToRem(72),
+    },
+    groupD: {
+      fontSize: pixelsToRem(96),
+      lineHeight: pixelsToRem(104),
+    },
+  },
+  royal: {
+    groupA: {
+      fontSize: pixelsToRem(40),
+      lineHeight: pixelsToRem(44),
+    },
+    groupB: {
+      fontSize: pixelsToRem(52),
+      lineHeight: pixelsToRem(60),
+    },
+    groupD: {
+      fontSize: pixelsToRem(76),
+      lineHeight: pixelsToRem(84),
+    },
+  },
+  foolscap: {
+    groupA: {
+      fontSize: pixelsToRem(32),
+      lineHeight: pixelsToRem(36),
+    },
+    groupB: {
+      fontSize: pixelsToRem(40),
+      lineHeight: pixelsToRem(44),
+    },
+    groupD: {
+      fontSize: pixelsToRem(56),
+      lineHeight: pixelsToRem(60),
+    },
+  },
+  canon: {
+    groupA: {
+      fontSize: pixelsToRem(28),
+      lineHeight: pixelsToRem(32),
+    },
+    groupB: {
+      fontSize: pixelsToRem(32),
+      lineHeight: pixelsToRem(36),
+    },
+    groupD: {
+      fontSize: pixelsToRem(44),
+      lineHeight: pixelsToRem(48),
+    },
+  },
+  trafalgar: {
+    groupA: {
+      fontSize: pixelsToRem(20),
+      lineHeight: pixelsToRem(24),
+    },
+    groupB: {
+      fontSize: pixelsToRem(24),
+      lineHeight: pixelsToRem(28),
+    },
+    groupD: {
+      fontSize: pixelsToRem(32),
+      lineHeight: pixelsToRem(36),
+    },
+  },
+  paragon: {
+    groupA: {
+      fontSize: pixelsToRem(20),
+      lineHeight: pixelsToRem(24),
+    },
+    groupB: {
+      fontSize: pixelsToRem(22),
+      lineHeight: pixelsToRem(26),
+    },
+    groupD: {
+      fontSize: pixelsToRem(28),
+      lineHeight: pixelsToRem(32),
+    },
+  },
+  doublePica: {
+    groupA: {
+      fontSize: pixelsToRem(18),
+      lineHeight: pixelsToRem(22),
+    },
+    groupB: {
+      fontSize: pixelsToRem(20),
+      lineHeight: pixelsToRem(24),
+    },
+    groupD: {
+      fontSize: pixelsToRem(24),
+      lineHeight: pixelsToRem(28),
+    },
+  },
+  greatPrimer: {
+    groupA: {
+      fontSize: pixelsToRem(18),
+      lineHeight: pixelsToRem(22),
+    },
+    groupB: {
+      fontSize: pixelsToRem(18),
+      lineHeight: pixelsToRem(22),
+    },
+    groupD: {
+      fontSize: pixelsToRem(20),
+      lineHeight: pixelsToRem(24),
+    },
+  },
+  bodyCopy: {
+    groupA: {
+      fontSize: pixelsToRem(15),
+      lineHeight: pixelsToRem(20),
+    },
+    groupB: {
+      fontSize: pixelsToRem(16),
+      lineHeight: pixelsToRem(22),
+    },
+    groupD: {
+      fontSize: pixelsToRem(16),
+      lineHeight: pixelsToRem(22),
+    },
+  },
+  pica: {
+    groupA: {
+      fontSize: pixelsToRem(15),
+      lineHeight: pixelsToRem(20),
+    },
+    groupB: {
+      fontSize: pixelsToRem(16),
+      lineHeight: pixelsToRem(20),
+    },
+    groupD: {
+      fontSize: pixelsToRem(16),
+      lineHeight: pixelsToRem(20),
+    },
+  },
+  longPrimer: {
+    groupA: {
+      fontSize: pixelsToRem(15),
+      lineHeight: pixelsToRem(18),
+    },
+    groupB: {
+      fontSize: pixelsToRem(15),
+      lineHeight: pixelsToRem(18),
+    },
+    groupD: {
+      fontSize: pixelsToRem(14),
+      lineHeight: pixelsToRem(18),
+    },
+  },
+  brevier: {
+    groupA: {
+      fontSize: pixelsToRem(14),
+      lineHeight: pixelsToRem(18),
+    },
+    groupB: {
+      fontSize: pixelsToRem(14),
+      lineHeight: pixelsToRem(18),
+    },
+    groupD: {
+      fontSize: pixelsToRem(13),
+      lineHeight: pixelsToRem(16),
+    },
+  },
+  minion: {
+    groupA: {
+      fontSize: pixelsToRem(12),
+      lineHeight: pixelsToRem(16),
+    },
+    groupB: {
+      fontSize: pixelsToRem(12),
+      lineHeight: pixelsToRem(16),
+    },
+    groupD: {
+      fontSize: pixelsToRem(12),
+      lineHeight: pixelsToRem(16),
+    },
+  },
+};
+
+export default latin;
