--- src/app/utilities/pixelsToRem/index.js
@@ -1,3 +1,0 @@
-const pixelsToRem = pixels => pixels / 16;
-
-export default pixelsToRem;
