+++ src/app/utilities/pixelsToRem/index.ts
@@ -1,0 +1,3 @@
+const pixelsToRem = (pixels: number): number => pixels / 16;
+
+export default pixelsToRem;
